% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vectorfile.R
\name{carti_download}
\alias{carti_download}
\alias{download_vectorfile_url_all}
\title{Download a series of files from cartiflette project.}
\usage{
carti_download(
  values = c("28", "29"),
  borders = "COMMUNE",
  vectorfile_format = "geojson",
  filter_by = "DEPARTEMENT",
  year = 2022,
  provider = "IGN",
  source = "EXPRESS-COG-CARTO-TERRITOIRE",
  crs = 4326
)

download_vectorfile_url_all(...)
}
\arguments{
\item{borders}{The administrative level for filtering the data, now including arrondissements. Default is "COMMUNE_ARRONDISSEMENT".}

\item{vectorfile_format}{The format of the vector file to download. Default is "geojson".}

\item{filter_by}{The filter criterion for the data. Default is "DEPARTEMENT".}

\item{year}{The year of the data to download. Default is 2022.}

\item{provider}{The data provider. Default is "IGN".}

\item{source}{The data source, adjusted to the new dataset family parameter. Default is "EXPRESS-COG-CARTO-TERRITOIRE".}

\item{crs}{The coordinate reference system (CRS) for the data. Default is 4326.}
}
\value{
An 'sf' data frame containing the downloaded geographical vector data.
}
\description{
Download a series of files from cartiflette project.
}
\note{
`download_vectorfile_url_all` is deprecated and will be removed in
 a future version
}
\examples{
df <- carti_download(
   crs = 4326,
   values = c("75", "92", "93", "94"),
   borders="COMMUNE_ARRONDISSEMENT",
   vectorfile_format="geojson",
   filter_by="DEPARTEMENT",
   source="EXPRESS-COG-CARTO-TERRITOIRE",
   year=2022)
plot(sf::st_geometry(df))

}
