% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{create_url_download}
\alias{create_url_download}
\title{Create a URL to download cartiflette dataset.}
\usage{
create_url_download(
  base_url = "https://minio.lab.sspcloud.fr/projet-cartiflette/production",
  year = 2022,
  borders = "COMMUNE_ARRONDISSEMENT",
  crs = 4326,
  vectorfile_format = "geojson",
  filter_by = "DEPARTEMENT",
  provider = "IGN",
  source = "EXPRESS-COG-CARTO-TERRITOIRE",
  value = 75
)
}
\arguments{
\item{base_url}{The base URL for the data repository. Default is the URL for the project.}

\item{year}{The year of the data to download. Default is 2022.}

\item{borders}{The administrative level for filtering the data, now including arrondissements. Default is "COMMUNE_ARRONDISSEMENT".}

\item{crs}{The coordinate reference system (CRS) for the data. Default is 4326.}

\item{vectorfile_format}{The format of the vector file to download. Default is "geojson".}

\item{filter_by}{The filter criterion for the data. Default is "DEPARTEMENT".}

\item{provider}{The data provider. Default is "IGN".}

\item{source}{The data source, adjusted to the new dataset family parameter. Default is "EXPRESS-COG-CARTO-TERRITOIRE".}

\item{value}{The value to filter the data by. Default is 75.}
}
\value{
A character string representing the generated download URL.
}
\description{
This function generates a URL for downloading geographical vector data
based on the provided parameters. The URL is constructed to match the
specific format required by the data repository.
}
\examples{
create_url_download()
create_url_download(year = 2021, borders = "REGION", crs = 4326, value = 33)

}
