% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vectorfile.R
\name{download_vectorfile_url_single}
\alias{download_vectorfile_url_single}
\title{Download and read a single geographical vector file from a generated URL.}
\usage{
download_vectorfile_url_single(
  value = 75,
  borders = "COMMUNE",
  vectorfile_format = "geojson",
  filter_by = "DEPARTEMENT",
  year = 2022,
  provider = "IGN",
  source = "EXPRESS-COG-CARTO-TERRITOIRE",
  crs = 4326
)
}
\arguments{
\item{value}{The value to filter the data by. Default is 75.}

\item{borders}{The administrative level for filtering the data, now including arrondissements. Default is "COMMUNE_ARRONDISSEMENT".}

\item{vectorfile_format}{The format of the vector file to download. Default is "geojson".}

\item{filter_by}{The filter criterion for the data. Default is "DEPARTEMENT".}

\item{year}{The year of the data to download. Default is 2022.}

\item{provider}{The data provider. Default is "IGN".}

\item{source}{The data source, adjusted to the new dataset family parameter. Default is "EXPRESS-COG-CARTO-TERRITOIRE".}

\item{crs}{The coordinate reference system (CRS) for the data. Default is 4326.}
}
\value{
An 'sf' data frame containing the downloaded geographical vector data.
}
\description{
This function generates a URL using specified parameters and downloads a single
geographical vector file from the URL. The downloaded file is then read into an
R data frame using the 'sf' package.
}
\examples{
download_vectorfile_url_single()
plot(sf::st_geometry(download_vectorfile_url_single()))
df <- download_vectorfile_url_single(filter_by = "REGION", year = 2022, value = 52, crs = 4326)
plot(sf::st_geometry(df))
}
\seealso{
\code{\link{create_url_download}} for generating the download URL.
}
